+++ lib/src/Utils/dui_widget_registry.dart
@@ -7,9 +7,11 @@ import '../core/builders/dui_animated_button_builder.dart';
 import '../core/builders/dui_app_bar_builder.dart';
 import '../core/builders/dui_avatar_builder.dart';
 import '../core/builders/dui_button_builder.dart';
+import '../core/builders/dui_calendar_builder.dart';
 import '../core/builders/dui_carousel_builder.dart';
 import '../core/builders/dui_checkbox_builder.dart';
 import '../core/builders/dui_circular_progress_indicator_builder.dart';
+import '../core/builders/dui_conditional_builder_builder.dart';
 import '../core/builders/dui_container2_builder.dart';
 import '../core/builders/dui_custom_widget_builder.dart';
 import '../core/builders/dui_drawer_builder.dart';
@@ -24,17 +26,22 @@ import '../core/builders/dui_icon_button_builder.dart';
 import '../core/builders/dui_image_builder.dart';
 import '../core/builders/dui_listview_builder.dart';
 import '../core/builders/dui_lottie_builder.dart';
+import '../core/builders/dui_paginated_listview_builder.dart';
+import '../core/builders/dui_pin_field_builder.dart';
+import '../core/builders/dui_refresh_indicator_builder.dart';
 import '../core/builders/dui_rich_text_builder.dart';
 import '../core/builders/dui_scaffold_builder.dart';
 import '../core/builders/dui_sized_box_builder.dart';
 import '../core/builders/dui_spacer_builder.dart';
 import '../core/builders/dui_stack_builder.dart';
 import '../core/builders/dui_stepper_builder.dart';
+import '../core/builders/dui_stream_builder.dart';
 import '../core/builders/dui_switch_builder.dart';
 import '../core/builders/dui_tab_view_builder.dart';
 import '../core/builders/dui_tab_view_item_builder.dart';
 import '../core/builders/dui_text_builder.dart';
 import '../core/builders/dui_text_form_field_builder.dart';
+import '../core/builders/dui_timer_builder.dart';
 import '../core/builders/dui_vertical_divider_builder.dart';
 import '../core/builders/dui_video_player_builder.dart';
 import '../core/builders/dui_webview_builder.dart';
@@ -122,6 +129,17 @@ class DUIWidgetRegistry {
     'digia/checkbox': withoutRegistry(DUICheckboxBuilder.create),
     'digia/switch': withoutRegistry(DUISwitchBuilder.create),
     'digia/animatedButton': withoutRegistry(DUIAnimatedButtonBuilder.create),
+    'digia/pinField': withoutRegistry(DUIPinFieldBuilder.create),
+    'digia/calendar': withoutRegistry(DUICalendarBuilder.create),
+    'digia/paginatedListView': (data, {registry}) =>
+        DUIPaginatedListViewBuilder.create(data),
+    'digia/timer': withoutRegistry(DUITimerBuilder.create),
+    'digia/conditionalBuilder': (data, {registry}) =>
+        DUIConditionalBuilderBuilder.create(data,
+            registry: DUIWidgetRegistry.shared),
+    'digia/streamBuilder': (data, {registry}) => DUIStreamBuilder(data: data),
+    'digia/refreshIndicator': (data, {registry}) =>
+        DUIRefreshIndicator(data: data)
   };
 
   static const DUIWidgetRegistry shared = DUIWidgetRegistry();