+++ lib/src/components/dui_tab_view/dui_tab_view_props.dart
@@ -15,8 +15,10 @@ class DUITabViewProps {
   final double? dividerHeight;
   final bool? isScrollable;
   final String? indicatorColor;
+  final String? indicatorSize;
   final double? viewportFraction;
   final DUIInsets? tabBarPadding;
+  final DUIInsets? labelPadding;
   @JsonKey(
       fromJson: DUITextStyle.fromJson,
       includeToJson: false,
@@ -30,6 +32,8 @@ class DUITabViewProps {
 
   DUITabViewProps(
       {this.hasTabs,
+      this.indicatorSize,
+      this.labelPadding,
       this.selectedLabelColor,
       this.selectedLabelStyle,
       this.unselectedLabelColor,