+++ lib/src/core/builders/dui_text_builder.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:widget_marquee/widget_marquee.dart';
 
 import '../../Utils/basic_shared_utils/dui_decoder.dart';
 import '../../Utils/util_functions.dart';
@@ -39,9 +40,18 @@ class _DUIText extends StatelessWidget {
     final style =
         toTextStyle(DUITextStyle.fromJson(props['textStyle']), context);
     final maxLines = eval<int>(props['maxLines'], context: context);
-    final overflow = DUIDecoder.toTextOverflow(props['overflow']);
     final textAlign = DUIDecoder.toTextAlign(props['alignment']);
-
+    if (props['overflow'] == 'marquee') {
+      return Marquee(
+        pause: Duration.zero,
+        delay: Duration.zero,
+        duration: const Duration(seconds: 11),
+        gap: 100,
+        child: Text(text.toString(),
+            style: style, maxLines: maxLines, textAlign: textAlign),
+      );
+    }
+    final overflow = DUIDecoder.toTextOverflow(props['overflow']);
     return Text(text.toString(),
         style: style,
         maxLines: maxLines,