+++ lib/src/core/scoped_values.dart
@@ -5,6 +5,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../digia_ui.dart';
 import '../Utils/basic_shared_utils/lodash.dart';
 import 'app_state_provider.dart';
+import 'bracket_scope_provider.dart';
 import 'indexed_item_provider.dart';
 import 'page/dui_page_bloc.dart';
 
@@ -44,7 +45,15 @@ ExprContext createScope(BuildContext context, ExprContext? localScope) {
       }) ??
       pageScope;
 
-  // Wrap the scope chain from above over localSCope
-  return ifNotNull(localScope, ((p0) => p0..appendEnclosing(indexScope))) ??
+  final bScope = ifNotNull(BracketScope.maybeOf(context), (p0) {
+        return ExprContext(
+            name: 'bracket',
+            variables:
+                Map.fromEntries(p0.variables.map((e) => MapEntry(e.$1, e.$2))),
+            enclosing: indexScope);
+      }) ??
       indexScope;
+
+  // Wrap the scope chain from above over localSCope
+  return ifNotNull(localScope, ((p0) => p0..appendEnclosing(bScope))) ?? bScope;
 }